Move saml2 rest package to match its new URI
