+++ synapse/rest/synapse/client/__init__.py
@@ -52,7 +52,7 @@ def build_synapse_client_resource_tree(hs: "HomeServer") -> Mapping[str, Resourc
         resources["/_synapse/client/oidc"] = OIDCResource(hs)
 
     if hs.config.saml2_enabled:
-        from synapse.rest.saml2 import SAML2Resource
+        from synapse.rest.synapse.client.saml2 import SAML2Resource
 
         res = SAML2Resource(hs)
         resources["/_synapse/client/saml2"] = res
