+++ synapse/rest/synapse/client/saml2/metadata_resource.py
